adbot-x-fullstack/backend/requirements.txt
fastapi==0.110.1
google-generativeai==0.8.6
httpx==0.28.1
edge-tts>=6.1.12,<8
pydantic==2.12.5
python-dotenv==1.2.1
uvicorn==0.25.0
websockets==15.0.1
